# proto/openapi.yaml
openapi: 3.1.0
info:
  title: SeaSnail OpenAPI
  version: 0.1.0
  description: |
    SeaSnail 本地音频转译服务的 HTTP/REST API。本地 UI 与第三方应用统一经此 API 接入（OpenAPI 为唯一集成面）。

    ## 鉴权
    Scoped Bearer Token，请求头 `Authorization: Bearer <token>`。token 独立于用户密码，可吊销、可多签、**绑 account_id**。
    可用 scope：
      - `sessions:read`    读取 / 搜索会话、取原始音频、导出
      - `sessions:write`  创建会话、编辑转译文本（**仅本地 UI token 可得，不对第三方签发**）
      - `sessions:delete` 删除会话（可按需签发给第三方）
      - `tokens:manage`    管理 token（**仅本地 UI token 可得**）

    **多账户模型**：MVP 支持多账户，每账户独立密码 + 独立 master DEK，账户间密钥完全隔离。日常免密只对**活跃账户**成立（其 DEK 缓存于 app 私有 keychain，`last_active_account_id` 记于非敏感 prefs 供重开取回）；切换账户需重新认证（`POST /accounts/{id}/unlock`）。

    - 首个账户：`POST /auth/setup`（免鉴权、仅 loopback、仅"未初始化"态可用）建首个账户并签发该账户 **root token**（`is_root=true` + `account_id`，持有全部 scope），root token secret 存 app 私有 keychain，本地 UI 据此调用。
    - 后续账户：`POST /accounts`（需任一已解锁账户的 root token）——无全局 root，靠"已解锁即可建新账户"（已确认）。
    - 切换/解锁：`POST /accounts/{id}/unlock`（免鉴权 + password）→ 解 wrap 该账户 DEK 入 keychain 作活跃 → 签发/恢复该账户 root token。
    - 第三方 token 经 `POST /tokens` 签发，**绑 account_id**（仅可访问该账户、且仅该账户已解锁时），**非 root 调用方不可被授予 `sessions:write` / `tokens:manage`**（daemon 硬约束），故第三方仅可得 `sessions:read` / `sessions:delete`。
    - 改密走 `POST /auth/password`（需 root token + 当前密码）。

    本地优先：服务监听本机；bootstrap 不走云端，token 由本地 UI 解锁后在 Integrations 面板生成。

    ## 数据安全
    入库数据（原始音频 + 转译内容）经加密落库（per-account 独立密钥）；读取 / 导出时由服务解密后返回。
...
    导出为敏感操作，需在请求体中提供用户密码再确认（作用于活跃账户）。
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ss_live_
  responses:
    BadRequest:
      description: 请求字段或语义无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 未携带或无效 token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 权限不足（缺 scope / 密码错误）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 当前资源状态与请求冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Gone:
      description: 一次性资源已消费、过期或不可再用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PayloadTooLarge:
      description: 请求体超过接口限制
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnprocessableEntity:
      description: 请求可解析，但内容或当前配置不满足接口契约
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Locked:
      description: 目标账户未解锁（DEK 不在内存，需先 POST /accounts/{id}/unlock）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: insufficient_scope }
            message: { type: string }
          required: [code, message]
      required: [error]
...
    DictionaryEntry:
      type: object
      description: 账户级 Dictionary 词条视图；不暴露 normalized_term、learning_event_id 或历史输入全文
      properties:
        id: { type: string, format: uuid }
        term: { type: string, description: 用户确认的规范写法（保留大小写与兼容字符形式） }
        source: { type: string, enum: [manual, learned], description: manual=用户明确维护；learned=自动学习获得 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, term, source, created_at, updated_at]
    DictionaryPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/DictionaryEntry' }
        next_cursor: { type: string, nullable: true, description: 不透明分页游标，透传回 cursor 查询参数；与当前 query 摘要绑定，末页为 null }
      required: [items]
    AddDictionaryEntriesRequest:
      type: object
      description: 单条或小批量增量添加；整请求校验，任一词条非法或写入后超账户容量则零写入
      properties:
        terms:
          type: array
          minItems: 1
          maxItems: 1000
          items: { type: string, minLength: 1, maxLength: 128, description: 单词条最多 128 个 Unicode scalar value 且 UTF-8 不超过 512 bytes }
      required: [terms]
      additionalProperties: false
    DictionaryMutationResult:
      type: object
      description: 增量写入统计；请求内重复及与既有 manual 冲突计入 skipped_count，与既有 learned 冲突提升为 manual
      properties:
        added:
          type: array
          items: { $ref: '#/components/schemas/DictionaryEntry' }
        promoted:
          type: array
          description: 由 learned 提升为 manual 的词条（learning_event_id 已清空，不再受撤销影响）
          items: { $ref: '#/components/schemas/DictionaryEntry' }
        skipped_count: { type: integer, minimum: 0 }
      required: [added, promoted, skipped_count]
    ImportDictionaryCsvRequest:
      type: object
      description: JSON 包裹的 UTF-8 CSV 文本；固定单列 term 表头，RFC 4180 字段引用与双引号转义，允许 UTF-8 BOM 与 LF/CRLF；解析或校验失败整次零写入
      properties:
        csv: { type: string, minLength: 1, maxLength: 1048576, description: CSV 全文（按 UTF-8 计最多 1 MiB）；解析后数据记录最多 1000 条 }
      required: [csv]
      additionalProperties: false
    DictionaryImportResult:
      allOf:
        - $ref: '#/components/schemas/DictionaryMutationResult'
        - type: object
          properties:
            parsed_count: { type: integer, minimum: 0, description: CSV 解析出的数据记录数（不含表头与空白物理行） }
          required: [parsed_count]
    DictionaryImportPreview:
      type: object
      description: |
        CSV 只读预览统计。valid_count 为通过词条校验并按 normalized_term 去重后的数量；
        added_count、promoted_count、skipped_count 基于预览时的当前词典计算，正式提交仍会在事务内重新校验。
      properties:
        parsed_count: { type: integer, minimum: 0, description: CSV 解析出的数据记录数（不含表头与空白物理行） }
        valid_count: { type: integer, minimum: 0, description: 校验通过且按 normalized_term 去重后的词条数 }
        added_count: { type: integer, minimum: 0, description: 按预览时状态预计新增的词条数 }
        promoted_count: { type: integer, minimum: 0, description: 按预览时状态预计由 learned 提升为 manual 的词条数 }
        skipped_count: { type: integer, minimum: 0, description: 请求内重复及与既有 manual 冲突而预计跳过的记录数 }
      required: [parsed_count, valid_count, added_count, promoted_count, skipped_count]
...
    LearningSubmission:
      type: object
      description: |
        消费 learning ticket 提交一次学习。mode=cleanup：候选完全取自 ticket 绑定的已验证 Cleanup corrections，candidates 必须省略或为空；
        mode=user_edit：candidates 为原生层从用户修改中提取的候选词条，daemon 侧再次执行统一词条校验。
      properties:
        ticket: { type: string, description: InjectionPlan 返回的 learning_ticket }
        mode: { type: string, enum: [cleanup, user_edit] }
        candidates:
          type: array
          maxItems: 32
          items: { type: string, minLength: 1, maxLength: 128 }
          description: 仅 mode=user_edit 时携带；mode=cleanup 时必须省略或为空
      required: [ticket, mode]
      additionalProperties: false
    LearningResult:
      type: object
      description: 学习写入结果；没有实际新增时 learning_event_id=null，调用方不显示成功通知
      properties:
        learning_event_id: { type: string, nullable: true, description: 本批学习事件 ID（撤销凭据）；仅在实际新增时返回 }
        added_terms:
          type: array
          items: { type: string }
          description: 本批真正新增的规范词条（此前已存在的不重复计入）
      required: [learning_event_id, added_terms]
    UndoLearningResult:
      type: object
      properties:
        undone_count: { type: integer, minimum: 0, description: 实际删除的词条数；重复撤销返回 0（幂等） }
      required: [undone_count]
security:
  - bearerAuth: []
paths:
...
  /internal/dictionary/learning-events:
    post:
      summary: 消费 learning ticket 并提交一次学习
      description: |
        Tauri 原生观察层专用（root token，禁止经 WebView 通用 API 代理）。ticket 一次性、短期有效（TTL 2 分钟），
        内存绑定账户/Session/InjectionPlan generation；账户已切换、Session 非已完成或 generation 不匹配返回 409
        （learning_ticket_mismatch），过期、已消费或因容量淘汰返回 410（learning_ticket_expired /
        learning_ticket_consumed），编码非法返回 400（learning_ticket_invalid）。账户锁定沿用全局 423（locked）
        映射；ticket 在账户校验前已原子消费，任何拒绝均不可重放。
        mode=cleanup 时候选完全取自 ticket 绑定的已验证 Cleanup corrections（仅 proper_noun/phonetic），
        附带 candidates 或 mode 非法返回 400；mode=user_edit 的 candidates 最多 32 项，由 daemon 再次执行
        统一词条校验，词条数量或内容不合法返回 422（dictionary_invalid_term）。
        同一 Session 的 cleanup 采纳与 user_edit 学习归并为单次学习结果。错误 message 不含词条、ticket 或目标输入框文本。
      x-internal: true
      x-required-root: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LearningSubmission' }
      responses:
        '200':
          description: 学习写入结果；无实际新增时 learning_event_id=null
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LearningResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/Gone' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '423': { $ref: '#/components/responses/Locked' }
  /internal/dictionary/learning-events/{id}/undo:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string }, description: LearningResult 返回的 learning_event_id }
    post:
      summary: 撤销一批学习新增
      description: |
        Tauri 原生层专用（root token，禁止经 WebView 通用 API 代理）。仅删除该事件真正新增且仍为 learned 的词条
        （期间被用户编辑或确认提升为 manual 的不受影响）。幂等：重复撤销返回 undone_count=0。
      x-internal: true
      x-required-root: true
      responses:
        '200':
          description: 撤销结果（幂等）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UndoLearningResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '423': { $ref: '#/components/responses/Locked' }
...
  /dictionary:
    get:
      summary: 搜索/分页列出当前账户的 Dictionary 词条
      description: |
        仅 root token。排序固定为 manual 在前、learned 在后，同来源按 updated_at DESC, id ASC。
        query 经统一规范化后对 normalized_term 做包含匹配（SQL LIKE 的 %、_ 与 escape 字符按普通字符处理）。
        cursor 绑定排序元组与 query 摘要：无法解码、与当前 query 不匹配或指向非法排序值时返回 400。
      x-required-root: true
      parameters:
        - { in: query, name: query, schema: { type: string, maxLength: 128 }, description: 可选；最多 128 个 Unicode scalar value }
        - { in: query, name: cursor, schema: { type: string }, description: 分页游标（上一页响应的 next_cursor 原样透传） }
        - { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '423': { $ref: '#/components/responses/Locked' }
  /dictionary/entries:
    post:
      summary: 增量添加词条（单条或小批量）
      description: |
        仅 root token。整请求校验：任一词条非法（空值/控制字符/纯标点空白/超长）或写入后将超过单账户容量
        （1,000 条）时返回 422 且零写入（dictionary_invalid_term / dictionary_limit_exceeded）。
        增量语义：请求内重复与既有 manual 词条计入 skipped_count；与既有 learned 词条冲突时更新其规范写法、
        提升为 manual 并清空 learning_event_id（计入 promoted）。新增、提升与去重在同一事务内完成，
        数据库 UNIQUE(normalized_term) 为最终并发去重边界；并发冲突折算到 skipped_count，不暴露 409。
      x-required-root: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddDictionaryEntriesRequest' }
      responses:
        '200':
          description: 增量写入统计（部分词条可能计入 promoted/skipped）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryMutationResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '423': { $ref: '#/components/responses/Locked' }
    delete:
      summary: 清空当前账户全部用户词条
      description: 仅 root token；幂等。删除全部 manual 与 learned 词条（未来系统保护词条不经本操作删除）。UI 必须先明确影响并二次确认。
      x-required-root: true
      responses:
        '204': { description: 已清空（幂等） }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '423': { $ref: '#/components/responses/Locked' }
  /dictionary/entries/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    put:
      summary: 编辑词条的规范写法
      description: |
        仅 root token。仅接受 {term}，由服务端执行权威规范化。与另一词条规范化冲突返回 409（dictionary_conflict）。
        编辑 learned 词条会将其提升为 manual 并清空 learning_event_id（不再受撤销影响）。
      x-required-root: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                term: { type: string, minLength: 1, maxLength: 128 }
              required: [term]
              additionalProperties: false
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '423': { $ref: '#/components/responses/Locked' }
    delete:
      summary: 删除单条词条
      description: 仅 root token；幂等，存在或不存在均返回 204。
      x-required-root: true
      responses:
        '204': { description: 已删除（幂等） }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '423': { $ref: '#/components/responses/Locked' }
  /dictionary/imports/preview:
    post:
      summary: 预览 CSV 增量导入结果
      description: |
        仅 root token。使用与正式导入相同的 CSV 解析、词条校验、规范化和当前词典状态，返回只读统计，
        不写入词条或学习事件。预览不锁定数据库状态；正式提交时必须重新校验并以提交结果为准。
      x-required-root: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportDictionaryCsvRequest' }
      responses:
        '200':
          description: 导入预览统计（解析/有效/预计新增/预计提升/预计跳过）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryImportPreview' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '423': { $ref: '#/components/responses/Locked' }
  /dictionary/imports:
    post:
      summary: CSV 增量导入词条
      description: |
        仅 root token。body 为 JSON 包裹的 UTF-8 CSV 文本：固定且仅允许单列 term 表头，RFC 4180 字段引用与双引号
        转义，允许 UTF-8 BOM 与 LF/CRLF，空白物理行忽略。CSV 语法错误、缺少/重复/未知表头、额外列、显式空 term、
        非法词条或超过 1,000 条数据记录时返回 422（dictionary_csv_invalid / dictionary_invalid_term），
        整次导入零写入。合法请求在一个事务内增量插入，不删除既有词条；合法重复记录按增量语义计入 skipped_count。
        客户端可先调用 /dictionary/imports/preview 展示预计统计；正式提交会重新校验，最终以本响应为准。
      x-required-root: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportDictionaryCsvRequest' }
      responses:
        '200':
          description: 导入统计（解析数/新增/提升/跳过）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DictionaryImportResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '423': { $ref: '#/components/responses/Locked' }
  /dictionary/export:
    get:
      summary: 导出当前账户全部用户词条为 CSV
      description: |
        Tauri 原生 command 专用（root token，禁止经 WebView 通用 API 代理）：原生层获取完整 CSV 后弹保存对话框
        写文件，文件路径不进入 daemon，导出不依赖 UI 当前分页状态。固定输出 term 表头 + RFC 4180 记录，
        不输出 BOM、内部 ID、来源、时间戳或事件 ID；顺序固定 normalized_term ASC, id ASC；
        导出内容可由 POST /dictionary/imports 无损重新导入。
      x-internal: true
      x-required-root: true
      responses:
        '200':
          description: "UTF-8 CSV（Content-Type: text/csv; charset=utf-8；下载文件名 seasnail-dictionary.csv）"
          content:
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '423': { $ref: '#/components/responses/Locked' }
...